Prevent duplicated login message
Entering raw mode during an interactive terminal session (like using a
remote shell) can print out the login message twice. This is caused by a
WINDOW_BUFFER_SIZE_EVENT (that we caused by changing console settings)
to be translated into a SIGWINCH, which is not handled, and thus treated
as an error by the write thread, after already successfully writing to
the console.