Reset the console code page before exiting
The console's input and output code pages may not have been properly
reset when the process exits.

First, the output page was being forced to UTF8 without the previous
value being saved. This can be removed since we are already setting the
codepage ourselves in a later function call.

Second, GetConsoleCP() and GetConsoleOutputCP() can fail and return 0,
leaving us without a saved version of the previous value if we then
immediately set the code page without checking the result of the above
functions. Also note that, because they are called after we have already
forced UTF8, we are incorrectly restoring to UTF8 even when the call to
GetConsoleOutputCP() succeeds.